try caching glob matches
